ref: Remove old OTel support code
