+++ scripts/populate_tox/tox.jinja
@@ -42,9 +42,6 @@ envlist =
     # GCP
     {py3.7}-gcp
 
-    # OpenTelemetry (OTel)
-    {py3.7,py3.9,py3.12,py3.13,py3.14,py3.14t}-opentelemetry
-
     # OpenTelemetry with OTLP
     {py3.7,py3.9,py3.12,py3.13,py3.14}-otlp
 
@@ -138,10 +135,6 @@ deps =
     aws_lambda: requests
     aws_lambda: uvicorn
     aws_lambda: pyyaml
-
-    # OpenTelemetry (OTel)
-    opentelemetry: opentelemetry-distro
-    opentelemetry: pytest-forked
 
     # OpenTelemetry with OTLP
     otlp: opentelemetry-distro[otlp]
@@ -240,7 +233,6 @@ setenv =
     aws_lambda: _TESTPATH=tests/integrations/aws_lambda
     cloud_resource_context: _TESTPATH=tests/integrations/cloud_resource_context
     gcp: _TESTPATH=tests/integrations/gcp
-    opentelemetry: _TESTPATH=tests/integrations/opentelemetry
     otlp: _TESTPATH=tests/integrations/otlp
     potel: _TESTPATH=tests/integrations/opentelemetry
     socket: _TESTPATH=tests/integrations/socket
