+++ scripts/split_tox_gh_actions/split_tox_gh_actions.py
@@ -163,7 +163,6 @@ GROUPS = {
     ],
     "Misc": [
         "loguru",
-        "opentelemetry",
         "otlp",
         "potel",
         "pure_eval",
