--- sentry_sdk/integrations/opentelemetry/__init__.py
@@ -1,7 +1,0 @@
-from sentry_sdk.integrations.opentelemetry.propagator import SentryPropagator
-from sentry_sdk.integrations.opentelemetry.span_processor import SentrySpanProcessor
-
-__all__ = [
-    "SentryPropagator",
-    "SentrySpanProcessor",
-]
